+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## 0.2.0 - Unreleased
 
+- Added the `app requests` command for fetching the log of requests made by an application.
 - Made the `app register` command to store application properties locally, so that they are easy to
   use with other commands.
-- Added the `app requests` command for fetching the log of requests made by an application.
+- Added optional parameters for the `app register` command: `-d/--description`, `--gdpr-email`,
+  `--privacy-url` and `--terms-url`.
 - Small fixes and improvements.
 
 ## 0.1.1 - 2024-01-10